Renamed plugin-api to messaging
